+++ data/src/main/java/com/lunchplay/data/database/MemoDatabase.kt
@@ -1,7 +1,9 @@
-package com.lunchplay.data
+package com.lunchplay.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.lunchplay.data.memo.source.local.MemoDao
+import com.lunchplay.data.memo.source.local.MemoEntity
 
 @Database(entities = [MemoEntity::class], version = 1)
 abstract class MemoDatabase : RoomDatabase() {
