+++ data/src/main/java/com/lunchplay/data/memo/MemoRepositoryImpl.kt
@@ -1,13 +1,14 @@
-package com.lunchplay.data
+package com.lunchplay.data.memo
 
-import com.lunchplay.domain.Memo
-import com.lunchplay.domain.MemoRepository
+import com.lunchplay.data.memo.source.local.MemoDataSource
+import com.lunchplay.domain.entity.Memo
+import com.lunchplay.domain.repository.MemoRepository
 import io.reactivex.Flowable
 import javax.inject.Inject
 
 class MemoRepositoryImpl @Inject constructor(
     private val memoDataSource: MemoDataSource
-) : MemoRepository{
+) : MemoRepository {
     override fun getMemos(): Flowable<List<Memo>> =
         memoDataSource.getMemos().flatMap { memoEntities ->
             Flowable.just(memoEntities.map { memoEntity ->
