+++ data/src/main/java/com/lunchplay/data/memo/source/local/MemoDao.kt
@@ -1,4 +1,4 @@
-package com.lunchplay.data
+package com.lunchplay.data.memo.source.local
 
 import androidx.room.Dao
 import androidx.room.Query
