+++ domain/src/main/java/com/lunchplay/domain/entity/Memo.kt
@@ -1,4 +1,4 @@
-package com.lunchplay.domain
+package com.lunchplay.domain.entity
 
 data class Memo(
     val id: Int,
