+++ domain/src/main/java/com/lunchplay/domain/repository/MemoRepository.kt
@@ -1,5 +1,6 @@
-package com.lunchplay.domain
+package com.lunchplay.domain.repository
 
+import com.lunchplay.domain.entity.Memo
 import io.reactivex.Flowable
 
 interface MemoRepository {
