+++ domain/src/main/java/com/lunchplay/domain/usecase/GetMemos.kt
@@ -1,5 +1,6 @@
-package com.lunchplay.domain
+package com.lunchplay.domain.usecase
 
+import com.lunchplay.domain.repository.MemoRepository
 import javax.inject.Inject
 
 class GetMemos @Inject constructor(
