+++ presentation/build.gradle
@@ -46,7 +46,6 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     //module
-    implementation project(path: ':data')
     implementation project(path: ':domain')
 
     //rxjava2
