+++ presentation/src/main/res/layout/activity_memo.xml
@@ -7,7 +7,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.lunchplay.ui.MemoActivity">
+        tools:context="com.lunchplay.ui.memo.MemoActivity">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView_memoList"
